feat: rename graph module to domain
