+++ gryf/src/domain/generic.rs
