+++ gryf/src/domain/path.rs
